adjust v1

+++ venda/backend/src/app/controllers/ClienteController.js
@@ -1,6 +1,6 @@
 import Cliente from '../models/Cliente';
 import ClienteService from '../services/ClienteService';
-
+import AppError from '../exception/AppError';
 export default class ClienteController {
 
 	constructor() {
@@ -16,7 +16,7 @@ export default class ClienteController {
 	async busca(req, res) {
 
 		const cliente = Cliente.create(req.body);
-
+		
 		const result = await this.clienteService.busca(cliente);
 
 		return res.json(result);
@@ -40,6 +40,11 @@ export default class ClienteController {
 	}
 
 	async update(req, res) {
+
+		if (!['ROLE_MASTER', 'ROLE_ADMIN'].includes(req.user.role)) {
+			throw new AppError("Usuário não autorizado!", 401);
+		}
+
 		const { id } = req.params;
 		const cliente = Cliente.create(req.body);
 
@@ -49,12 +54,15 @@ export default class ClienteController {
 	}
 
 	async delete(req, res) {
+
+		if (!['ROLE_MASTER', 'ROLE_ADMIN'].includes(req.user.role)) {
+			throw new AppError("Usuário não autorizado!", 401);
+		}
+
 		const { id } = req.params;
 		await this.clienteService.delete(id);
 		return res.status(204).json();
 	}
-
-
 
 
 }
