+++ venda/backend/src/app/controllers/PedidoItemController.js
@@ -1,5 +1,6 @@
 import PedidoItem from '../models/PedidoItem';
 import PedidoItemService from '../services/PedidoItemService';
+import AppError from '../exception/AppError';
 
 export default class PedidoItemController {
 
@@ -16,6 +17,10 @@ export default class PedidoItemController {
 
 	async create(req, res) {
 
+		if (!['ROLE_MASTER', 'ROLE_ADMIN'].includes(req.user.role)) {
+			throw new AppError("Usuário não autorizado!", 401);
+		}
+
 		const pedidoItem = PedidoItem.create(req.body);
 
 		const result = await this.pedidoItemService.insert(pedidoItem);
@@ -25,6 +30,11 @@ export default class PedidoItemController {
 	}
 
 	async delete(req, res) {
+
+		if (!['ROLE_MASTER', 'ROLE_ADMIN'].includes(req.user.role)) {
+			throw new AppError("Usuário não autorizado!", 401);
+		}
+		
 		const { id } = req.params;
 		await this.pedidoItemService.delete(id);
 		return res.status(204).json();
